[jsscripting] Move `TimerImpl` inside `ThreadsafeTimers`
